+++ frontend/plugin/src/lib/components/generate-document-configuration/generate-document-config.util.ts
@@ -40,7 +40,6 @@ export function isGenerateDocumentConfigValid(
     options.selectedCatalogId &&
     config &&
     config.templateId &&
-    config.outputFormat &&
     options.filename &&
     options.filename.trim() &&
     isProcessVariableNameValid(config.resultProcessVariable)